airflow/requirements.txt
# Airflow providers
apache-airflow-providers-postgres==5.10.0

# Cosmos for dbt integration
astronomer-cosmos==1.3.0

# dbt adapter for ClickHouse
dbt-clickhouse==1.7.2

# Additional utilities
sqlalchemy==1.4.51
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.2.4
